refactor(auth): Refactor authorizers  (#469)
* Move auth in their own folder

* refactor(auth): Move relevant functions to ProposalAuth

* refactor(auth): update ProposalAuthorization signature

* refactor(auth): Split questionary authorizers

* fix(auth): Resolve merge conflict

* fix(auth): Rephrazing authorizer
