+++ src/auth/ProposalAuthorization.ts
@@ -1,51 +1,50 @@
-import { inject, injectable } from 'tsyringe';
+import { container, inject, injectable } from 'tsyringe';
 
 import { Tokens } from '../config/Tokens';
+import { CallDataSource } from '../datasources/CallDataSource';
+import { ProposalDataSource } from '../datasources/ProposalDataSource';
 import { ReviewDataSource } from '../datasources/ReviewDataSource';
 import { SEPDataSource } from '../datasources/SEPDataSource';
-import { UserDataSource } from '../datasources/UserDataSource';
-import { Proposal } from '../models/Proposal';
-import { Roles } from '../models/Role';
+import { VisitDataSource } from '../datasources/VisitDataSource';
+import { ProposalStatusDefaultShortCodes } from '../models/ProposalStatus';
 import { User, UserWithRole } from '../models/User';
-import { ProposalDataSource } from './../datasources/ProposalDataSource';
-import { VisitDataSource } from './../datasources/VisitDataSource';
+import { Proposal } from '../resolvers/types/Proposal';
+import { ProposalSettingsDataSource } from './../datasources/ProposalSettingsDataSource';
+import { UserDataSource } from './../datasources/UserDataSource';
+import { UserAuthorization } from './UserAuthorization';
 
 @injectable()
-export class UserAuthorization {
+export class ProposalAuthorization {
+  private userAuth = container.resolve(UserAuthorization);
   constructor(
-    @inject(Tokens.UserDataSource) private userDataSource: UserDataSource,
-    @inject(Tokens.ReviewDataSource) private reviewDataSource: ReviewDataSource,
-    @inject(Tokens.SEPDataSource) private sepDataSource: SEPDataSource,
-    @inject(Tokens.VisitDataSource) private visitDataSource: VisitDataSource,
     @inject(Tokens.ProposalDataSource)
-    private proposalDataSource: ProposalDataSource
+    private proposalDataSource: ProposalDataSource,
+    @inject(Tokens.UserDataSource)
+    private userDataSource: UserDataSource,
+    @inject(Tokens.ReviewDataSource)
+    private reviewDataSource: ReviewDataSource,
+    @inject(Tokens.SEPDataSource)
+    private sepDataSource: SEPDataSource,
+    @inject(Tokens.VisitDataSource)
+    private visitDataSource: VisitDataSource,
+    @inject(Tokens.CallDataSource)
+    private callDataSource: CallDataSource,
+    @inject(Tokens.ProposalSettingsDataSource)
+    private proposalSettingsDataSource: ProposalSettingsDataSource
   ) {}
 
-  // User officer has access
-  isUserOfficer(agent: UserWithRole | null) {
-    if (agent == null) {
-      return false;
+  private async resolveProposal(
+    proposalOrProposalId: Proposal | number
+  ): Promise<Proposal | null> {
+    let proposal;
+
+    if (typeof proposalOrProposalId === 'number') {
+      proposal = await this.proposalDataSource.get(proposalOrProposalId);
+    } else {
+      proposal = proposalOrProposalId;
     }
 
-    return agent?.currentRole?.shortCode === Roles.USER_OFFICER;
-  }
-
-  isUser(agent: UserWithRole | null) {
-    if (agent == null) {
-      return false;
-    }
-
-    return agent?.currentRole?.shortCode === Roles.USER;
-  }
-
-  async hasRole(agent: UserWithRole | null, role: string): Promise<boolean> {
-    if (agent == null) {
-      return false;
-    }
-
-    return this.userDataSource.getUserRoles(agent.id).then((roles) => {
-      return roles.some((roleItem) => roleItem.shortCode === role);
-    });
+    return proposal;
   }
 
   isPrincipalInvestigatorOfProposal(
@@ -58,14 +57,6 @@ export class UserAuthorization {
     if (agent.id === proposal.proposerId) {
       return true;
     }
-  }
-
-  async isInstrumentScientist(agent: UserWithRole) {
-    if (agent == null) {
-      return false;
-    }
-
-    return agent?.currentRole?.shortCode === Roles.INSTRUMENT_SCIENTIST;
   }
 
   async isMemberOfProposal(
@@ -142,36 +133,33 @@ export class UserAuthorization {
       });
   }
 
-  async isSampleSafetyReviewer(agent: UserWithRole | null) {
-    if (agent == null) {
+  isVisitorOfProposal(
+    agent: UserWithRole,
+    proposalPk: number
+  ): boolean | PromiseLike<boolean> {
+    return this.visitDataSource.isVisitorOfProposal(agent.id, proposalPk);
+  }
+
+  async isChairOrSecretaryOfProposal(agent: User | null, proposalPk: number) {
+    if (agent == null || !agent.id || !proposalPk) {
       return false;
     }
 
-    return agent?.currentRole?.shortCode === Roles.SAMPLE_SAFETY_REVIEWER;
+    return this.sepDataSource.isChairOrSecretaryOfProposal(
+      agent.id,
+      proposalPk
+    );
   }
 
-  private async resolveProposal(
-    proposalOrProposalId: Proposal | number
-  ): Promise<Proposal | null> {
-    let proposal;
-
-    if (typeof proposalOrProposalId === 'number') {
-      proposal = await this.proposalDataSource.get(proposalOrProposalId);
-    } else {
-      proposal = proposalOrProposalId;
-    }
-
-    return proposal;
-  }
-  async hasAccessRights(
+  async hasReadRights(
     agent: UserWithRole | null,
     proposal: Proposal
   ): Promise<boolean>;
-  async hasAccessRights(
+  async hasReadRights(
     agent: UserWithRole | null,
     proposalId: number
   ): Promise<boolean>;
-  async hasAccessRights(
+  async hasReadRights(
     agent: UserWithRole | null,
     proposalOrProposalId: Proposal | number
   ): Promise<boolean> {
@@ -186,9 +174,9 @@ export class UserAuthorization {
     }
 
     return (
-      this.isUserOfficer(agent) ||
-      this.hasGetAccessByToken(agent) ||
-      (await this.isInstrumentScientist(agent)) ||
+      this.userAuth.isUserOfficer(agent) ||
+      this.userAuth.isInstrumentScientist(agent) ||
+      this.userAuth.hasGetAccessByToken(agent) ||
       (await this.isMemberOfProposal(agent, proposal)) ||
       (await this.isReviewerOfProposal(agent, proposal.primaryKey)) ||
       (await this.isScientistToProposal(agent, proposal.primaryKey)) ||
@@ -196,53 +184,58 @@ export class UserAuthorization {
       (await this.isVisitorOfProposal(agent, proposal.primaryKey))
     );
   }
-  isVisitorOfProposal(
-    agent: UserWithRole,
-    proposalPk: number
-  ): boolean | PromiseLike<boolean> {
-    return this.visitDataSource.isVisitorOfProposal(agent.id, proposalPk);
-  }
 
-  async isChairOrSecretaryOfSEP(
-    agent: User | null,
-    sepId: number
-  ): Promise<boolean> {
-    if (agent == null || !agent.id || !sepId) {
-      return false;
-    }
+  private async isProposalEditable(proposal: Proposal): Promise<boolean> {
+    const proposalStatus =
+      await this.proposalSettingsDataSource.getProposalStatus(
+        proposal.statusId
+      );
 
-    return this.sepDataSource.isChairOrSecretaryOfSEP(agent.id, sepId);
-  }
-
-  async isChairOrSecretaryOfProposal(agent: User | null, proposalPk: number) {
-    if (agent == null || !agent.id || !proposalPk) {
-      return false;
-    }
-
-    return this.sepDataSource.isChairOrSecretaryOfProposal(
-      agent.id,
-      proposalPk
+    return (
+      proposal.submitted === false ||
+      proposalStatus?.shortCode ===
+        ProposalStatusDefaultShortCodes.EDITABLE_SUBMITTED
     );
   }
 
-  hasGetAccessByToken(agent: UserWithRole) {
-    return !!agent.accessPermissions?.['ProposalQueries.get'];
-  }
-
-  async isMemberOfSEP(
+  async hasWriteRights(
     agent: UserWithRole | null,
-    sepId: number
+    proposal: Proposal
+  ): Promise<boolean>;
+  async hasWriteRights(
+    agent: UserWithRole | null,
+    proposalId: number
+  ): Promise<boolean>;
+  async hasWriteRights(
+    agent: UserWithRole | null,
+    proposalOrProposalId: Proposal | number
   ): Promise<boolean> {
-    if (agent == null || !agent.currentRole) {
+    if (!agent) {
       return false;
     }
 
-    const [sep] = await this.sepDataSource.getUserSepsByRoleAndSepId(
-      agent.id,
-      agent.currentRole,
-      sepId
-    );
+    const proposal = await this.resolveProposal(proposalOrProposalId);
 
-    return sep !== null;
+    if (!proposal) {
+      return false;
+    }
+
+    if (
+      this.userAuth.isUserOfficer(agent) ||
+      this.userAuth.hasGetAccessByToken(agent)
+    ) {
+      return true;
+    }
+
+    const callId = proposal.callId;
+    const isMemberOfProposal = await this.isMemberOfProposal(agent, proposal);
+    const isCallActive = await this.callDataSource.checkActiveCall(callId);
+    const isProposalEditable = await this.isProposalEditable(proposal);
+
+    if (isMemberOfProposal && isCallActive && isProposalEditable) {
+      return true;
+    }
+
+    return false;
   }
 }
