+++ src/auth/ShipmentAuthorization.ts
@@ -4,11 +4,11 @@ import { Tokens } from '../config/Tokens';
 import { ShipmentDataSource } from '../datasources/ShipmentDataSource';
 import { UserWithRole } from '../models/User';
 import { Shipment } from './../resolvers/types/Shipment';
-import { UserAuthorization } from './UserAuthorization';
+import { ProposalAuthorization } from './ProposalAuthorization';
 
 @injectable()
 export class ShipmentAuthorization {
-  private userAuth = container.resolve(UserAuthorization);
+  private proposalAuth = container.resolve(ProposalAuthorization);
   constructor(
     @inject(Tokens.ShipmentDataSource)
     private shipmentDataSource: ShipmentDataSource
@@ -64,11 +64,6 @@ export class ShipmentAuthorization {
     agent: UserWithRole | null,
     shipmentOrShipmentId: Shipment | number
   ) {
-    // User officer has read/write rights
-    if (this.userAuth.isUserOfficer(agent)) {
-      return true;
-    }
-
     const shipment = await this.resolveShipment(shipmentOrShipmentId);
     if (!shipment) {
       return false;
@@ -76,8 +71,8 @@ export class ShipmentAuthorization {
 
     /*
      * For the shipment the authorization follows the business logic for the proposal
-     * authorization that the shipment is associated with
+     * authorization that the shipment is associated with, this should be changed
      */
-    return this.userAuth.hasAccessRights(agent, shipment.proposalPk);
+    return this.proposalAuth.hasReadRights(agent, shipment.proposalPk);
   }
 }
