+++ src/auth/questionary/ShipmentDeclarationQuestionaryAuthorizer.ts
@@ -1,0 +1,68 @@
+import { logger } from '@esss-swap/duo-logger';
+import { container, inject, injectable } from 'tsyringe';
+
+import { Tokens } from '../../config/Tokens';
+import { ProposalDataSource } from '../../datasources/ProposalDataSource';
+import { ShipmentDataSource } from '../../datasources/ShipmentDataSource';
+import { UserWithRole } from '../../models/User';
+import { ProposalAuthorization } from '../ProposalAuthorization';
+import { QuestionaryAuthorizer } from '../QuestionaryAuthorization';
+import { UserAuthorization } from '../UserAuthorization';
+
+@injectable()
+export class ShipmentDeclarationQuestionaryAuthorizer
+  implements QuestionaryAuthorizer
+{
+  private userAuth = container.resolve(UserAuthorization);
+  private proposalAuth = container.resolve(ProposalAuthorization);
+
+  constructor(
+    @inject(Tokens.ProposalDataSource)
+    private proposalDataSource: ProposalDataSource,
+    @inject(Tokens.ShipmentDataSource)
+    private shipmentDataSource: ShipmentDataSource
+  ) {}
+  async hasReadRights(agent: UserWithRole | null, questionaryId: number) {
+    return this.hasRights(agent, questionaryId);
+  }
+  async hasWriteRights(agent: UserWithRole | null, questionaryId: number) {
+    return this.hasRights(agent, questionaryId);
+  }
+
+  private async hasRights(agent: UserWithRole | null, questionaryId: number) {
+    if (!agent) {
+      return false;
+    }
+
+    if (this.userAuth.isUserOfficer(agent)) {
+      return true;
+    }
+
+    const queryResult = await this.shipmentDataSource.getShipments({
+      filter: { questionaryIds: [questionaryId] },
+    });
+
+    if (queryResult.length !== 1) {
+      logger.logError(
+        'Expected to find exactly one sample with questionaryId',
+        { questionaryId }
+      );
+
+      return false;
+    }
+
+    const shipment = queryResult[0];
+
+    const proposal = await this.proposalDataSource.get(shipment.proposalPk);
+
+    if (!proposal) {
+      logger.logError('Could not find proposal for questionary', {
+        questionaryId,
+      });
+
+      return false;
+    }
+
+    return this.proposalAuth.hasReadRights(agent, proposal);
+  }
+}
