+++ src/mutations/VisitMutations.ts
@@ -1,5 +1,6 @@
 import { container, inject, injectable } from 'tsyringe';
 
+import { VisitAuthorization } from '../auth/VisitAuthorization';
 import { Tokens } from '../config/Tokens';
 import { ProposalDataSource } from '../datasources/ProposalDataSource';
 import { QuestionaryDataSource } from '../datasources/QuestionaryDataSource';
@@ -17,12 +18,13 @@ import { VisitRegistration } from '../models/VisitRegistration';
 import { CreateVisitArgs } from '../resolvers/mutations/CreateVisitMutation';
 import { UpdateVisitArgs } from '../resolvers/mutations/UpdateVisitMutation';
 import { UpdateVisitRegistrationArgs } from '../resolvers/mutations/UpdateVisitRegistration';
-import { UserAuthorization } from '../utils/UserAuthorization';
-import { VisitAuthorization } from './../utils/VisitAuthorization';
+import { ProposalAuthorization } from './../auth/ProposalAuthorization';
+import { UserAuthorization } from './../auth/UserAuthorization';
 
 @injectable()
 export default class VisitMutations {
   private userAuth = container.resolve(UserAuthorization);
+  private proposalAuth = container.resolve(ProposalAuthorization);
   private visitAuth = container.resolve(VisitAuthorization);
 
   constructor(
@@ -87,7 +89,7 @@ export default class VisitMutations {
 
     if (proposal === null) {
       return rejection(
-        'Can not create visit, proposal for the scheduled does not exist',
+        'Can not create visit, proposal for the scheduled event does not exist',
         {
           args,
           agent: user,
@@ -108,7 +110,10 @@ export default class VisitMutations {
       );
     }
 
-    const isProposalOwner = await this.userAuth.hasAccessRights(user, proposal);
+    const isProposalOwner = await this.proposalAuth.hasReadRights(
+      user,
+      proposal
+    );
     if (isProposalOwner === false) {
       return rejection(
         'Can not create visit for proposal that does not belong to you',
