+++ src/queries/ShipmentQueries.ts
@@ -1,13 +1,13 @@
 import { logger } from '@esss-swap/duo-logger';
 import { container, inject, injectable } from 'tsyringe';
 
+import { ShipmentAuthorization } from '../auth/ShipmentAuthorization';
 import { Tokens } from '../config/Tokens';
 import { ShipmentDataSource } from '../datasources/ShipmentDataSource';
 import { Authorized } from '../decorators';
 import { Roles } from '../models/Role';
 import { UserWithRole } from '../models/User';
 import { ShipmentsArgs } from '../resolvers/queries/ShipmentsQuery';
-import { ShipmentAuthorization } from '../utils/ShipmentAuthorization';
 
 @injectable()
 export default class ShipmentQueries {
